fix: rename constructor slice
