+++ src/components/burger-ingredient/burger-ingredient.tsx
@@ -3,7 +3,7 @@ import { useLocation } from 'react-router-dom';
 import { BurgerIngredientUI } from '@ui';
 import { TBurgerIngredientProps } from './type';
 import { useAppDispatch } from '../../services/store';
-import { addIngredient } from '../../services/slices/constructorSlice';
+import { addIngredient } from '../../services/slices/constructSlice';
 
 export const BurgerIngredient: FC<TBurgerIngredientProps> = memo(
   ({ ingredient, count }) => {
